:recycle: refactor: pool card component
